fix: JsonPrettyPrinter component
